+++ crates/sml-lower/src/dec.rs
@@ -558,6 +558,9 @@ fn get_one(cx: &mut Cx, dec: ast::DecOne) -> sml_hir::DecIdx {
             .collect();
           // not the greatest, since we have no body at all if the ptrs are None. but if they were
           // both None, then something's very strange about the fun_bind_cases anyway.
+          if let Some(0) = num_pats {
+            cx.err(dec.syntax().text_range(), ErrorKind::EmptyFun);
+          }
           let exp = {
             let arg_names: Vec<_> = (0..num_pats.unwrap_or(1)).map(|_| cx.fresh()).collect();
             let mut arg_exprs =